move chips helper to common
